% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_LIV.R
\name{LIV}
\alias{LIV}
\title{Function to run simple Latent Instrument Variable model via MLE}
\usage{
LIV(formula, endogenous.regressor, data, iters = NULL, parameters = c())
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted. Of class "formula".}

\item{endogenous.regressor}{str the name of the endogenous variable in the input data}

\item{data}{A data.frame containing the data of all parts specified in the formula parameter.}

\item{iters}{number of iterations for the optimization.}

\item{parameters}{A named vector containing a set of parameters to use in the first optimization iteration.
The names have to correspond exactly to the names of the components specified in the formula parameter.
If not provided, a linear model is fitted to derive them.}
}
\value{
A named vector of all coefficients resulting from model fitting.
}
\description{
Fits linear models with one endogenous regressor and no additional explanatory variables using the latent instrumental variable approach
}
\examples{
\donttest{
data("dataLatentIV")

# function call without any initial parameter values
livcoef <- LIV(y~x, "x", dataLatentIV, NULL)
livcoef


# function call with initial parameter values given by the user
fit<-kmeans(dataLatentIV$x, 2)
starts<-c(1, 2.5, fit$centers[1], fit$centers[2], 0.5, 1, 0.5, 1)
livcoef2<-LIV(y~x, "x", dataLatentIV, NULL, starts)
livcoef2

}

}
\references{
Ebbes, P., Wedel,M., Böckenholt, U., and Steerneman, A. G. M. (2005). 'Solving and Testing for Regressor-Error
(in)Dependence When no Instrumental Variables are Available: With New Evidence for the Effect of Education on Income'.
Quantitative Marketing and Economics, 3:365--392.

Ebbes P., Wedel M., Böckenholt U. (2009). “Frugal IV Alternatives to Identify the Parameter for an Endogenous Regressor.” Journal of Applied Econometrics, 24(3), 446–468.
}
\seealso{
\code{\link[optimx]{optimx}} for possible elements of parameter \code{optimx.arg}
}
