% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_LIV_LL.R
\name{LIV_LL}
\alias{LIV_LL}
\title{Latent Instrumental Variables (LIV) Log-likelihood function to optimize}
\usage{
LIV_LL(starting.parameters, data)
}
\arguments{
\item{starting.parameters}{the 8 parameters ("Intercept", "Endogenous regressor", "pi1", "pi2", "lambda1",
"Var(E)", "Cov(EV)", "Var(V)") to estimate as outlined in Ebbes 2004}

\item{data}{the input data.frame}
}
\value{
The log-likelihood function
}
\description{
A helper function to construct the log-likelihood function for simple LIV model
}
